Fix a number of issues on OSX 10.5
There still are some test failures in pyobjc-core and I haven't tested
the framework wrappers yet. The failures in core are fairly obscure, and
might even be due to platform limitations.

+++ pyobjc-core/Modules/objc/objc_support.m
@@ -2902,7 +2902,7 @@ depythonify_c_value(const char *type, PyObject *argument, void *datum)
             if (tmp != NULL) {
                 double dblval = PyFloat_AsDouble(tmp);
                 Py_DECREF(tmp);
-                *(float*) datum = dblval;
+                *(float*) datum = (float)dblval;
                 return 0;
             }
